requirements.txt
Flask==3.0.0
tensorflow==2.19.0
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.26.0
nltk==3.8.1
seaborn==0.13.0
matplotlib==3.8.0
swifter==1.3.4
tweepy==4.12.0
snscrape==0.7.0
streamlit
transformers
torch
python-dotenv
